% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orders.R
\name{archived_orders}
\alias{archived_orders}
\title{Get archived orders}
\usage{
archived_orders(base, quote)
}
\arguments{
\item{base}{Base currency}

\item{quote}{Quote currency}
}
\value{
A tibble
}
\description{
Get archived orders
}
\examples{
\dontrun{
archived_orders()
}
}
