% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency-limits.R
\name{currency_limits}
\alias{currency_limits}
\title{Get currency limits for all pairs}
\usage{
currency_limits()
}
\value{
A tibble
}
\description{
Get currency limits for all pairs
}
\examples{
currency_limits()
}
