% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-key.R
\name{get_api_nonce}
\alias{get_api_nonce}
\title{Get nonce}
\usage{
get_api_nonce(digits = 14)
}
\arguments{
\item{digits}{Number of digits in nonce string.}
}
\value{
A character vector.
}
\description{
If you change the way that the nonce is generated then you might need to
regenerate your API key. The reason for this is that the nonce associated
with any given key can only increase.
}
\examples{
get_api_nonce()
}
