% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-key.R
\name{get_api_secret}
\alias{get_api_secret}
\title{Get API secret}
\usage{
get_api_secret()
}
\value{
The API secret.
}
\description{
Get API secret
}
\examples{
set_credentials(secret = Sys.getenv("CEX_API_SECRET"))
get_api_secret()
}
