% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-key.R
\name{get_api_signature}
\alias{get_api_signature}
\title{Get API signature}
\usage{
get_api_signature()
}
\value{
A character vector.
}
\description{
Get API signature
}
\examples{
\dontrun{
get_api_signature()
}
}
