% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-key.R
\name{get_api_user_id}
\alias{get_api_user_id}
\title{Get User ID}
\usage{
get_api_user_id()
}
\value{
The user ID.
}
\description{
Get User ID
}
\examples{
set_credentials(user_id = Sys.getenv("CEX_USER_ID"))
get_api_user_id()
}
