% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orders.R
\name{get_order}
\alias{get_order}
\title{Get order details}
\usage{
get_order(order_id)
}
\arguments{
\item{order_id}{Order ID.}
}
\value{
A data frame.
}
\description{
Get order details
}
\examples{
\dontrun{
get_order()
}
}
