% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{get_order_tx}
\alias{get_order_tx}
\title{Get order transactions}
\usage{
get_order_tx(order_id)
}
\arguments{
\item{transaction_id}{Transaction ID}
}
\value{
A tibble
}
\description{
Get order transactions
}
\examples{
\dontrun{
get_order_tx()
}
}
