% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price.R
\name{last_price}
\alias{last_price}
\title{Get last price}
\usage{
last_price(base, quote)
}
\arguments{
\item{base}{Base currency}

\item{quote}{Quote currency}
}
\value{
A tibble
}
\description{
Wraps \code{GET https://cex.io/api/last_price/{symbol1}/{symbol2}}.
}
\examples{
last_price("ETH", "USD")
last_price("BTC", "USD")
}
