% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-key.R
\name{set_credentials}
\alias{set_credentials}
\title{Set API credentials}
\usage{
set_credentials(user_id, key, secret)
}
\arguments{
\item{api_key}{A CEX API key}

\item{api_secret}{A CEX API secret}
}
\description{
Set API credentials
}
\examples{
CEX_API_USER_ID  = Sys.getenv("CEX_USER_ID")
CEX_API_KEY      = Sys.getenv("CEX_API_KEY")
CEX_API_SECRET   = Sys.getenv("CEX_API_SECRET")
set_api_key(CEX_API_KEY, CEX_API_SECRET)
}
