% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticker.R
\name{ticker}
\alias{ticker}
\title{Get ticker}
\usage{
ticker(base, quote)
}
\arguments{
\item{base}{Base currency}

\item{quote}{Quote currency}
}
\value{
A tibble
}
\description{
Wraps \code{GET https://cex.io/api/ticker/{symbol1}/{symbol2}}.
}
\examples{
ticker("ETH", "USD")
}
