% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticker.R
\name{tickers}
\alias{tickers}
\title{Get tickers}
\usage{
tickers(symbols)
}
\arguments{
\item{symbols}{A vector of ticker symbols to be used as quote currency in pairs}
}
\value{
A tibble
}
\description{
Get tickers for a selection of quote currencies.
}
\details{
Wraps \code{https://cex.io/api/tickers/{symbol1}/{symbol2}/.../{symbolN}}.
}
\examples{
# All pairs with USD as quote.
tickers("USD")
# All pairs with ETH or BTC as quote.
tickers(c("ETH", "BTC"))
}
