% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-key.R
\name{get_api_nonce}
\alias{get_api_nonce}
\title{Get nonce}
\usage{
get_api_nonce(digits = 11)
}
\arguments{
\item{digits}{Number of digits in nonce string.}
}
\value{
A character vector.
}
\description{
Get nonce
}
\examples{
get_api_nonce()
}
